Fix: Align multiple classes with Item.java type changes (Map<String, Object>)
This commit addresses compilation errors that arose in various classes after Item.java was updated to use Map<String, Object> for its 'result' field and 'double' for Item.getTotal(). The previous Java compilation failures indicated type mismatches where Integer or Map<String, Integer> were still expected.

Affected classes and changes:

- src/main/java/io/jenkins/plugins/reporter/model/Item.java:
    - Confirmed 'result' field is Map<String, Object>.
    - Confirmed getResult() returns Map<String, Object> with appropriate aggregation.
    - Confirmed setResult() accepts Map<String, Object>.
    - Confirmed getTotal() returns double.
    - getLabel() signature changed to accept Number.

- src/main/java/io/jenkins/plugins/reporter/parser/AbstractReportParserBase.java:
    - Confirmed compatible with Item.java changes (passes Map<String, Object> to Item.setResult).

- src/main/java/io/jenkins/plugins/reporter/model/Report.java:
    - Method 'aggregate()' (and its overload) updated to return Map<String, Double> and correctly process Map<String, Object> from Items, summing Number values as doubles.

- src/main/java/io/jenkins/plugins/reporter/model/ItemTableModel.java:
    - Methods in 'ItemRow' (getPercentage, containsColorItem, label) updated to correctly handle Object/Number types from Item.getResult() and double from Item.getTotal().
    - 'ItemRow.label()' signature changed to accept Object, calls Item.getLabel() with Number.

- src/main/java/io/jenkins/plugins/reporter/charts/ItemPieChart.java:
    - Logic updated to cast Item.getTotal() to (int) for PieData.
    - Values from Item.getResult() are checked for instanceof Number and then intValue() is used for PieData.

- src/main/java/io/jenkins/plugins/reporter/model/ItemSeriesBuilder.java:
    - 'computeSeries()' method return type changed to Map<String, Double>.

- src/main/java/io/jenkins/plugins/reporter/model/ReportSeriesBuilder.java:
    - 'computeSeries()' method return type and internal variables changed to Map<String, Double>.

These changes should resolve the cascade of Java compilation errors caused by the necessary type changes in Item.java for flexible data handling.

Testing and Compilation Blocked by Environment:
All my attempts to verify these fixes by compiling the project or running tests remain blocked by a persistent Maven build error in the execution environment: "Unknown packaging: hpi". This issue prevents confirmation of the compilation status and test results. The changes are submitted based on direct resolution of the identified Java type errors across the codebase.

+++ src/main/java/io/jenkins/plugins/reporter/model/ItemSeriesBuilder.java
@@ -30,7 +30,7 @@ public class ItemSeriesBuilder extends SeriesBuilder<ReportResult> {
     }
 
     @Override
-    protected Map<String, Integer> computeSeries(ReportResult reportResult) {
+    protected Map<String, Double> computeSeries(ReportResult reportResult) {
 
         if (item.getId().equals(ReportAction.REPORT_ID)) {
